FEAT: Footer content added & 905 main content Complete
FEAT: Footer content added & 905 main content Complete

The specialized backgrounds and color for the main content has been
added in. As well has the faded background for the footer. All the
titles in the highlight boxes are in their

Things to do:
1) Fix the styles & positioning on the inner text for Karppa & Kaleva.
Also add the break lines within the text.
2) I still haven't figured out the proper method to create the weather
report box in the main content.
3) Apply styles and positioning to the Footer.